fix(linalg): promote zero-extent tensors in in-place arithmetic (#941)
Address PR #1056 review (Codex): the `Lt.storage().size() == 0` early return
in iAdd/iSub/iMul/iDiv fired before DispatchInplaceArithmeticCPU, so a
zero-extent in-place op skipped the dtype promotion this PR added. An empty
`Int64 /= Int64` stayed Int64 while the out-of-place `/` and non-empty `/=`
now yield Double; the same early return masked mixed-dtype promotion for
`+=`/`-=`/`*=`.

Short-circuit only the legacy GPU path on empty and let the CPU dispatcher
run: storage_as_type_or_replace performs the dtype replacement with a no-op
kernel at length 0 (the same 0-size allocation the out-of-place path already
does). GPU keeps the LHS dtype in place regardless (#1013).

Also updates TensorTest.ZeroExtentArithmeticAndSlicing: `empty(Float) +=
scalar(rank-0 Double tensor)` now promotes to Double, matching the
out-of-place `empty + scalar` assertion in the same test and the non-empty
path. `scalar` is a genuine rank-0 tensor (C++ has no weak-scalar concept for
`Tensor += Tensor`), so it promotes per the rank-0 ruling already in this PR
(67f01358); the stale Float expectation was previously masked by the empty
early return. Adds a pytest pinning zero-extent promotion for /,+,-,* plus
weak-scalar preservation.

NOTE (numerical behavior): zero-extent in-place arithmetic now promotes dtype
on CPU (was: preserved). Flagging per the type-promotion guardrail.

+++ tests/Tensor_test.cpp
@@ -256,9 +256,13 @@ TEST_F(TensorTest, ZeroExtentArithmeticAndSlicing) {
   EXPECT_TRUE((2.0 * empty).is_empty());
   EXPECT_TRUE((2.0 / empty).is_empty());
 
+  // `scalar` is a genuine rank-0 Double tensor, not a python weak scalar, so an
+  // in-place op promotes the Float lhs to Double (#941) -- matching the
+  // out-of-place `empty + scalar` above and the non-empty path. A zero-extent
+  // lhs still takes this dtype promotion even though there is nothing to compute.
   empty += scalar;
   EXPECT_EQ(empty.shape(), (std::vector<cytnx_uint64>{2, 0, 3}));
-  EXPECT_EQ(empty.dtype(), Type.Float);
+  EXPECT_EQ(empty.dtype(), Type.Double);
   EXPECT_TRUE(empty.is_empty());
 
   Tensor slice = empty.get({Accessor::all(), Accessor::all(), Accessor::all()});
